# back-end/swagger.yaml
openapi: 3.0.3
info:
  title: APIs tisco - OpenAPI 3.0
  description: APIs for tisco
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
  - url: http://159.138.242.188:3000/api/v1
...
components:
  securitySchemes:
    beareAuth:
        type: http
        scheme: bearer
        name: Authorization
        in: header
        bearerFormat: JWT
        description: >-
          Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345"
...
paths:
...
  /term:
    get:
      tags:
        - term
      summary: Get term and condition
      operationId: getTermAndCondition
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  term:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...